Refactor minimal API endpoints into dedicated modules
