+++ backend/Program.cs
@@ -1,10 +1,10 @@
 using backend.Clients;
 using backend.Data;
+using backend.Endpoints;
 using backend.Models;
 using backend.Observability;
 using backend.Policies;
 using backend.Services;
-using System.Security.Claims;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using OpenTelemetry.Trace;
@@ -69,89 +69,9 @@ var app = builder.Build();
 app.UseAuthentication();
 app.UseAuthorization();
 
-app.MapPost("/api/auth/login", async (LoginRequest request, UserService userService) =>
-{
-    var success = await userService.LoginAsync(request);
-    return success ? Results.Ok() : Results.Unauthorized();
-}).AllowAnonymous();
-
-app.MapPost("/api/auth/logout", async (UserService userService) =>
-{
-    await userService.LogoutAsync();
-    return Results.Ok();
-});
-
-app.MapPost("/api/auth/change-password", async (ChangePasswordRequest request, UserManager<AppUser> userManager, UserService userService, ClaimsPrincipal user) =>
-{
-    var appUser = await userManager.GetUserAsync(user);
-    if (appUser is null) return Results.Unauthorized();
-    var success = await userService.ChangePasswordAsync(appUser, request);
-    return success ? Results.Ok() : Results.BadRequest();
-}).RequireAuthorization();
-
-app.MapGet("/api/items", (int page = 1, int pageSize = 25, string? search = "") =>
-{
-    var allItems = Enumerable.Range(1, 100)
-        .Select(i => new { id = i, name = $"Item {i}" });
-
-    if (!string.IsNullOrWhiteSpace(search))
-    {
-        allItems = allItems.Where(i => i.name.Contains(search, StringComparison.OrdinalIgnoreCase));
-    }
-
-    var total = allItems.Count();
-    var items = allItems
-        .Skip((page - 1) * pageSize)
-        .Take(pageSize)
-        .ToList();
-
-    return Results.Ok(new { items, totalCount = total });
-});
-
-app.MapGet("/api/promocodes", (string? promoCode = "", string? status = "", int page = 1, int pageSize = 10) =>
-{
-    var allCodes = Enumerable.Range(1, 50).Select(i => new PromoCode
-    {
-        PromoCode = $"CODE{i:D3}",
-        Variance = "Active",
-        Description = $"Promo code {i}",
-        GeneratedOn = DateTime.UtcNow.AddDays(-i),
-        GeneratedBy = "System",
-        ValidFrom = DateTime.UtcNow.AddDays(-i),
-        ValidTo = DateTime.UtcNow.AddDays(i),
-        Status = i % 2 == 0 ? "Active" : "Inactive"
-    });
-
-    if (!string.IsNullOrWhiteSpace(promoCode))
-    {
-        allCodes = allCodes.Where(c => c.PromoCode.Equals(promoCode, StringComparison.OrdinalIgnoreCase));
-    }
-    if (!string.IsNullOrWhiteSpace(status))
-    {
-        allCodes = allCodes.Where(c => c.Status.Equals(status, StringComparison.OrdinalIgnoreCase));
-    }
-
-    var totalItems = allCodes.Count();
-    var data = allCodes.Skip((page - 1) * pageSize).Take(pageSize).ToList();
-    var totalPages = (int)Math.Ceiling(totalItems / (double)pageSize);
-
-    return Results.Ok(new
-    {
-        Data = data,
-        Pagination = new
-        {
-            CurrentPage = page,
-            PageSize = pageSize,
-            TotalPages = totalPages,
-            TotalItems = totalItems
-        }
-    });
-});
-
-app.MapGet("/products", async (IProductsClient client, HttpContext ctx) =>
-{
-    var data = await client.GetProductsAsync(ctx.RequestAborted);
-    return Results.Ok(data);
-}).RequireAuthorization();
+app.MapAuthEndpoints();
+app.MapItemEndpoints();
+app.MapPromoCodeEndpoints();
+app.MapProductEndpoints();
 
 app.Run();
